Refactor item-cost-analysis into its own file
This is to allow other code paths (besides the FeedValues tab) to use
this information.

Note that we are also now computing (and caching) the cost of every
product in the game, rather than just the science packs.
